Braian Ramirez
adición del paso a paso de la receta a ingresar por teclado, eliminación de parámetros incompletos

+++ Application/Application/Racion.cs
@@ -10,9 +10,9 @@ namespace Application
     internal class Racion
     {
         private string _producto { get; set; }
-        private int _peso { get; set; }
+        private double _peso { get; set; }
         public string Producto { get {return _producto;} set { _producto = value; } }
-        public int Peso { get { return _peso;} set {  _peso = value; } }
+        public double Peso { get { return _peso;} set {  _peso = value; } }
 
         public Racion(string producto, int peso)
         {
